vlm/requirements.txt
ollama
Pillow
PyYAML
pyrealsense2
numpy
opencv-python
